refactor: separated cli and command functions.
